requirements.txt
numpy
rawpy
matplotlib
scipy
PyQt6
opencv-python-headless
tifffile
pygame
pillow
